% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2tpm.R
\name{counts2tpm}
\alias{counts2tpm}
\title{Convert counts to transcripts per million (TPM).}
\usage{
counts2tpm(DGEobject, featureLength, meanFragmentLength)
}
\arguments{
\item{DGEobject}{A DGE List that contains raw feature counts (i.e.
n. fragments assigned to each gene), in DGE$counts.}

\item{featureLength}{A numeric vector with feature lengths.}

\item{meanFragmentLength}{A numeric vector with mean fragment lengths.}
}
\value{
tpm A numeric matrix normalized by library size and feature length.
}
\description{
Convert a numeric matrix of features (rows) and conditions (columns) with
raw feature counts to transcripts per million.
}
\details{
Lior Pachter. Models for transcript quantification from RNA-Seq.
   arXiv:1104.3889v2

Wagner, et al. Measurement of mRNA abundance using RNA-seq data:
   RPKM measure is inconsistent among samples. Theory Biosci. 24 July 2012.
   doi:10.1007/s12064-012-0162-3
}
