% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgeExtract.R
\name{dgeExtract}
\alias{dgeExtract}
\title{Extract tables in DGE objects (used in limma, edgeR and voom) into tidy data frames (tidyverse).}
\usage{
dgeExtract(object, datatype)
}
\arguments{
\item{object}{A DGE object - either a DGEList (edgeR) or an Elist (voom).}

\item{datatype}{Desired data table to extract - "counts", counts per million ("cpm"), "samples" or "genes"; "E", "weight"s, or "targets".}
}
\value{
A data_frame (aka tibble).
}
\description{
Extract tables in DGE objects (used in limma, edgeR and voom) into tidy data frames (tidyverse).
}
\examples{
dgeExtract(dummyDGEList,"counts")
dgeExtract(dummyDGEList,"samples")
}
