% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO2genes.R
\name{GO2genes}
\alias{GO2genes}
\title{Extract from a DGE object genes associated with a GO term of interest, e.g. from GSEA testing. Requires GOsource() output.}
\usage{
GO2genes(DGEobject, GO_query, resultsType)
}
\arguments{
\item{DGEobject}{A TopTags (DGE results) object. GeneID must be entrezgene, symbol or RefSeqID. Later versions will support ensemble_gene IDs and extract associated genes from a DGEList.}

\item{GO_query}{A GO term e.g. "GO:0002283"}

\item{resultsType}{Type of output to match. "results" indicates that DGEobject is a TopTags object. Future versions will allow "counts", or "cpm" output.}
}
\value{
A data_frame (aka tibble) to the console. EG.GO and entrez_joinTable are exported to R_GlobalEnv.
}
\description{
Extract from a DGE object genes associated with a GO term of interest, e.g. from GSEA testing. Requires GOsource() output.
}
\examples{
GO2genes(tt,"GO:0002283","results")
}
