% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOsource.R
\name{GOsource}
\alias{GOsource}
\title{Source a join table that enables linking entrezgene IDs with gene symbols (ALIAS) and accession numbers (ACCNUM).}
\usage{
GOsource(species)
}
\arguments{
\item{species}{A species identifier ("Hs", or "Mm")}
}
\value{
Creates a join table for the required species, which is called by GO2geneID().
}
\description{
Source a join table that enables linking entrezgene IDs with gene symbols (ALIAS) and accession numbers (ACCNUM).
}
