% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsExtract.R
\name{resultsExtract}
\alias{resultsExtract}
\title{Extract tables in DGE objects (used in limma, edgeR and voom) into tidy data frames (tidyverse).}
\usage{
resultsExtract(object)
}
\arguments{
\item{object}{An object of class TopTags}
}
\value{
A data_frame (aka tibble).
}
\description{
Extract tables in DGE objects (used in limma, edgeR and voom) into tidy data frames (tidyverse).
}
