% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableTOdf.R
\name{tableTOdf}
\alias{tableTOdf}
\title{Convert a table to tidy data_frame}
\usage{
tableTOdf(mytable)
}
\arguments{
\item{table}{A data.frame or matrix, or table nexted within a list}
}
\value{
A tidy data_frame (aka tibble), with rownames converted a to field named 'key'.
}
\description{
Convert a table to tidy data_frame
}
\examples{
tableTOdf(mtcars)
tableTOdf(NList$c)
}
